auth/requirements.txt
requests==2.22.0
fastapi==0.68.0
typing==3.7.4.1
pydantic==1.8.2
uvicorn==0.11.7
starlette==0.14.2
validators==0.18.2
SQLAlchemy==1.4.32
PyMySQL==1.0.2
PyJWT==1.7.0
bcrypt==3.2.0
pytest==7.0.1
